+++ tests/pytest/test_thermodynamics.py
@@ -1,4 +1,4 @@
-import fv3gfs.util
+import pace.util
 import numpy as np
 
 import fv3gfs.wrapper
@@ -10,7 +10,7 @@ class MockFv3GFS:
     def __init__(self):
         shape = [1, 1, 1]
         one = np.ones(shape, dtype=np.float32)
-        dims = [fv3gfs.util.Z_DIM, fv3gfs.util.Y_DIM, fv3gfs.util.X_DIM]
+        dims = [pace.util.Z_DIM, pace.util.Y_DIM, pace.util.X_DIM]
 
         self.dims = dims
         self.one = one
@@ -19,11 +19,11 @@ class MockFv3GFS:
     def get_state(self, names):
 
         state = {
-            "delp": fv3gfs.util.Quantity(self.one, units="Pa", dims=self.dims),
-            "specific_humidity": fv3gfs.util.Quantity(
+            "delp": pace.util.Quantity(self.one, units="Pa", dims=self.dims),
+            "specific_humidity": pace.util.Quantity(
                 self.one, units="", dims=self.dims
             ),
-            "cloud_water_mixing_ratio": fv3gfs.util.Quantity(
+            "cloud_water_mixing_ratio": pace.util.Quantity(
                 self.one, units="", dims=self.dims
             ),
         }
@@ -42,7 +42,7 @@ class MockFv3GFS:
 def test_set_state_mass_conserving_non_water():
     mock = MockFv3GFS()
     fv3gfs.wrapper.set_state_mass_conserving(
-        {"air_temperature": fv3gfs.util.Quantity(mock.one, dims=mock.dims, units="K")},
+        {"air_temperature": pace.util.Quantity(mock.one, dims=mock.dims, units="K")},
         fv3gfs=mock,
         pressure="delp",
     )
@@ -54,7 +54,7 @@ def test_set_state_mass_conserving_cant_set_delp():
     mock = MockFv3GFS()
     with pytest.raises(ValueError):
         fv3gfs.wrapper.set_state_mass_conserving(
-            {"delp": fv3gfs.util.Quantity(mock.one, dims=mock.dims, units="K")},
+            {"delp": pace.util.Quantity(mock.one, dims=mock.dims, units="K")},
             fv3gfs=mock,
             pressure="delp",
         )
@@ -64,7 +64,7 @@ def test_set_state_mass_conserving_water_added():
     mock = MockFv3GFS()
     fv3gfs.wrapper.set_state_mass_conserving(
         {
-            "specific_humidity": fv3gfs.util.Quantity(
+            "specific_humidity": pace.util.Quantity(
                 2 * mock.one, dims=mock.dims, units=""
             )
         },