+++ tests/test_set_ocean_surface_temperature.py
@@ -2,7 +2,7 @@ import unittest
 import os
 import numpy as np
 import fv3gfs.wrapper
-import fv3gfs.util
+import pace.util
 from mpi4py import MPI
 from util import (
     get_default_config,